apps/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
motor>=3.3.0
beanie>=1.19.0
tensorflow
tensorflow-hub>=0.16.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
pytest>=7.4.0
httpx>=0.24.0
pandas>=2.0.0
scipy>=1.11.0
email-validator>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
